moves emd class i/o into class files
